+++ packages/cloud-cognitive/src/components/EmptyStates/assets/NoTagsIllustration.js
@@ -12,11 +12,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import { pkg } from '../../../settings';
+import uuidv4 from '../../../global/js/utils/uuidv4';
 
 // The block part of our conventional BEM class names (blockClass__E--M).
 const blockClass = `${pkg.prefix}--empty-state`;
 
 export const NoTagsIllustration = ({ theme, size, ...rest }) => {
+  const svgId = uuidv4();
   return theme === 'dark' ? (
     <svg
       {...rest}
@@ -32,7 +34,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
     >
       <defs>
         <linearGradient
-          id="prefix__c"
+          id={`prefix__c_dark_${svgId}`}
           x1={34.96}
           y1={5.37}
           x2={48.48}
@@ -43,7 +45,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#393939" />
         </linearGradient>
         <linearGradient
-          id="prefix__a"
+          id={`prefix__a_dark_${svgId}`}
           x1={45.47}
           y1={46.54}
           x2={55.21}
@@ -54,15 +56,15 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#393939" />
         </linearGradient>
         <linearGradient
-          id="prefix__d"
+          id={`prefix__d_dark_${svgId}`}
           x1={38.25}
           y1={70.94}
           x2={38.25}
           y2={-3.31}
-          xlinkHref="#prefix__a"
+          xlinkHref={`#prefix__a_dark_${svgId}`}
         />
         <linearGradient
-          id="prefix__e"
+          id={`prefix__e_dark_${svgId}`}
           x1={28.77}
           y1={9.54}
           x2={48.15}
@@ -79,7 +81,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={0.73} stopColor="#525252" stopOpacity={0} />
         </linearGradient>
         <linearGradient
-          id="prefix__b"
+          id={`prefix__b_dark_${svgId}`}
           x1={51.28}
           y1={68.32}
           x2={51.34}
@@ -97,23 +99,23 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#d0d0d0" stopOpacity={0} />
         </linearGradient>
         <linearGradient
-          id="prefix__f"
+          id={`prefix__f_dark_${svgId}`}
           x1={51.09}
           y1={68.25}
           x2={51.22}
           y2={68.33}
-          xlinkHref="#prefix__b"
+          xlinkHref={`#prefix__b_dark_${svgId}`}
         />
         <linearGradient
-          id="prefix__g"
+          id={`prefix__g_dark_${svgId}`}
           x1={51.41}
           y1={68.32}
           x2={51.46}
           y2={68.35}
-          xlinkHref="#prefix__b"
+          xlinkHref={`#prefix__b_dark_${svgId}`}
         />
         <linearGradient
-          id="prefix__h"
+          id={`prefix__h_dark_${svgId}`}
           x1={33.94}
           y1={119.46}
           x2={42.79}
@@ -125,7 +127,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#161616" />
         </linearGradient>
         <linearGradient
-          id="prefix__i"
+          id={`prefix__i_dark_${svgId}`}
           x1={35.3}
           y1={23.46}
           x2={38.85}
@@ -146,33 +148,33 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
       />
       <path
         d="M45.47 30.38l6.58-3.76 3.12-1.85v.15a1.28 1.28 0 00-.17-.47L42.19 2.21c-.24-.42-.63-.65-.88-.51L28.47 9.12a.44.44 0 00-.18.4l13.88 3a1.38 1.38 0 00.62 1.07z"
-        fill="url(#prefix__c)"
+        fill={`url(#prefix__c_dark_${svgId})`}
       />
       <path
         d="M55 66.29a.46.46 0 00.18-.41V25.06a1 1 0 000-.29l-3.12 1.85-6.58 3.76 6.06 37.92z"
-        fill="url(#prefix__a)"
+        fill={`url(#prefix__a_dark_${svgId})`}
       />
       <path
         d="M51.53 26.47L38.69 4.23c-.25-.42-.64-.64-.88-.5L25 11.14a.43.43 0 00-.18.4v40.82a1.35 1.35 0 00.62 1.07l25.65 14.83c.34.2.62 0 .62-.36V27.08a1.33 1.33 0 00-.18-.61zM37.08 24a6.93 6.93 0 01-3.14-5.44c0-2 1.4-2.81 3.14-1.81a7 7 0 013.13 5.44c0 2.01-1.4 2.81-3.13 1.81z"
-        fill="url(#prefix__d)"
+        fill={`url(#prefix__d_dark_${svgId})`}
       />
       <path
         d="M39.05 4a1.71 1.71 0 00-.64-.64l-.6.34a.3.3 0 01.16 0 1 1 0 01.72.55l12.84 22.22a1.33 1.33 0 01.18.61V67.9a.41.41 0 01-.19.4l.58-.34V27.08a1.75 1.75 0 00-.24-.82z"
-        fill="url(#prefix__e)"
+        fill={`url(#prefix__e_dark_${svgId})`}
       />
-      <path d="M51.28 68.33z" fill="url(#prefix__b)" />
+      <path d="M51.28 68.33z" fill={`url(#prefix__b_dark_${svgId})`} />
       <path
         d="M51.23 68.32a.42.42 0 01-.14-.06.42.42 0 00.14.06z"
-        fill="url(#prefix__f)"
+        fill={`url(#prefix__f_dark_${svgId})`}
       />
-      <path d="M51.48 68.32h-.08z" fill="url(#prefix__g)" />
+      <path d="M51.48 68.32h-.08z" fill={`url(#prefix__g_dark_${svgId})`} />
       <path
         d="M39.66 22.17a6.93 6.93 0 01-3.14-5.44 1.36 1.36 0 010-.21c-1.48-.56-2.6.27-2.6 2.05A6.93 6.93 0 0037.08 24c1.66 1 3 .24 3.11-1.6a3.49 3.49 0 01-.53-.23z"
-        fill="url(#prefix__h)"
+        fill={`url(#prefix__h_dark_${svgId})`}
       />
       <path
         d="M35.67 16.33a2.88 2.88 0 011.41.43 7 7 0 013.13 5.44c0 1.42-.7 2.24-1.73 2.24a2.82 2.82 0 01-1.4-.43 6.93 6.93 0 01-3.14-5.44c0-1.42.7-2.24 1.73-2.24m0-.41c-1.3 0-2.15 1-2.15 2.65a7.3 7.3 0 003.35 5.8 3.23 3.23 0 001.61.48c1.3 0 2.15-1 2.15-2.65a7.29 7.29 0 00-3.35-5.8 3.2 3.2 0 00-1.61-.48z"
-        fill="url(#prefix__i)"
+        fill={`url(#prefix__i_dark_${svgId})`}
       />
     </svg>
   ) : (
@@ -190,7 +192,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
     >
       <defs>
         <linearGradient
-          id="prefix__b_dark"
+          id={`prefix__b_${svgId}`}
           x1={634.81}
           y1={667.18}
           x2={644.85}
@@ -202,7 +204,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopOpacity={0.1} />
         </linearGradient>
         <linearGradient
-          id="prefix__c_dark"
+          id={`prefix__c_${svgId}`}
           x1={29.21}
           y1={7.75}
           x2={56.14}
@@ -214,7 +216,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#c6c6c6" />
         </linearGradient>
         <linearGradient
-          id="prefix__d_dark"
+          id={`prefix__d_${svgId}`}
           x1={45.47}
           y1={46.54}
           x2={55.21}
@@ -225,7 +227,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#c6c6c6" />
         </linearGradient>
         <linearGradient
-          id="prefix__e_dark"
+          id={`prefix__e_${svgId}`}
           x1={38.25}
           y1={59.31}
           x2={38.25}
@@ -236,7 +238,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={0.78} stopColor="#e0e0e0" />
         </linearGradient>
         <linearGradient
-          id="prefix__f_dark"
+          id={`prefix__f_${svgId}`}
           x1={28.77}
           y1={9.54}
           x2={48.15}
@@ -256,7 +258,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={0.73} stopColor="#d0d0d0" stopOpacity={0} />
         </linearGradient>
         <linearGradient
-          id="prefix__a_dark"
+          id={`prefix__a_${svgId}`}
           x1={51.28}
           y1={68.32}
           x2={51.34}
@@ -274,23 +276,23 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#d0d0d0" stopOpacity={0} />
         </linearGradient>
         <linearGradient
-          id="prefix__g_dark"
+          id={`prefix__g_${svgId}`}
           x1={51.09}
           y1={68.25}
           x2={51.22}
           y2={68.33}
-          xlinkHref="#prefix__a_dark"
+          xlinkHref={`#prefix__a_${svgId}`}
         />
         <linearGradient
-          id="prefix__h_dark"
+          id={`prefix__h_${svgId}`}
           x1={51.41}
           y1={68.32}
           x2={51.46}
           y2={68.35}
-          xlinkHref="#prefix__a_dark"
+          xlinkHref={`#prefix__a_${svgId}`}
         />
         <linearGradient
-          id="prefix__i_dark"
+          id={`prefix__i_${svgId}`}
           x1={33.94}
           y1={19.46}
           x2={42.79}
@@ -301,7 +303,7 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
           <stop offset={1} stopColor="#8c8c8c" />
         </linearGradient>
         <linearGradient
-          id="prefix__j_dark"
+          id={`prefix__j_${svgId}`}
           x1={35.3}
           y1={23.46}
           x2={38.85}
@@ -319,38 +321,38 @@ export const NoTagsIllustration = ({ theme, size, ...rest }) => {
       </defs>
       <path fill="none" d="M0 0h80v80H0z" />
       <path
-        fill="url(#prefix__b_dark)"
+        fill={`url(#prefix__b_${svgId})`}
         d="M51.59 78.34l-26.8-15.48 3.62-2.09 26.8 15.48-3.62 2.09z"
       />
       <path
         d="M45.47 30.38l6.58-3.76 3.12-1.85v.15a1.28 1.28 0 00-.17-.47L42.19 2.21c-.24-.42-.63-.65-.88-.51L28.47 9.12a.44.44 0 00-.18.4l13.88 3a1.38 1.38 0 00.62 1.07z"
-        fill="url(#prefix__c_dark)"
+        fill={`url(#prefix__c_${svgId})`}
       />
       <path
         d="M55 66.29a.46.46 0 00.18-.41V25.06a1 1 0 000-.29l-3.12 1.85-6.58 3.76 6.06 37.92z"
-        fill="url(#prefix__d_dark)"
+        fill={`url(#prefix__d_${svgId})`}
       />
       <path
         d="M51.53 26.47L38.69 4.23c-.25-.42-.64-.64-.88-.5L25 11.14a.43.43 0 00-.18.4v40.82a1.35 1.35 0 00.62 1.07l25.65 14.83c.34.2.62 0 .62-.36V27.08a1.33 1.33 0 00-.18-.61zM37.08 24a6.93 6.93 0 01-3.14-5.44c0-2 1.4-2.81 3.14-1.81a7 7 0 013.13 5.44c0 2.01-1.4 2.81-3.13 1.81z"
-        fill="url(#prefix__e_dark)"
+        fill={`url(#prefix__e_${svgId})`}
       />
       <path
         d="M39.05 4a1.71 1.71 0 00-.64-.64l-.6.34a.3.3 0 01.16 0 1 1 0 01.72.55l12.84 22.22a1.33 1.33 0 01.18.61V67.9a.41.41 0 01-.19.4l.58-.34V27.08a1.75 1.75 0 00-.24-.82z"
-        fill="url(#prefix__f_dark)"
+        fill={`url(#prefix__f_${svgId})`}
       />
-      <path d="M51.28 68.33z" fill="url(#prefix__a_dark)" />
+      <path d="M51.28 68.33z" fill={`url(#prefix__a_${svgId})`} />
       <path
         d="M51.23 68.32a.42.42 0 01-.14-.06.42.42 0 00.14.06z"
-        fill="url(#prefix__g_dark)"
+        fill={`url(#prefix__g_${svgId})`}
       />
-      <path d="M51.48 68.32h-.08z" fill="url(#prefix__h_dark)" />
+      <path d="M51.48 68.32h-.08z" fill={`url(#prefix__h_${svgId})`} />
       <path
         d="M39.66 22.17a6.93 6.93 0 01-3.14-5.44 1.36 1.36 0 010-.21c-1.48-.56-2.6.27-2.6 2.05A6.93 6.93 0 0037.08 24c1.66 1 3 .24 3.11-1.6a3.49 3.49 0 01-.53-.23z"
-        fill="url(#prefix__i_dark)"
+        fill={`url(#prefix__i_${svgId})`}
       />
       <path
         d="M35.67 16.33a2.88 2.88 0 011.41.43 7 7 0 013.13 5.44c0 1.42-.7 2.24-1.73 2.24a2.82 2.82 0 01-1.4-.43 6.93 6.93 0 01-3.14-5.44c0-1.42.7-2.24 1.73-2.24m0-.41c-1.3 0-2.15 1-2.15 2.65a7.3 7.3 0 003.35 5.8 3.23 3.23 0 001.61.48c1.3 0 2.15-1 2.15-2.65a7.29 7.29 0 00-3.35-5.8 3.2 3.2 0 00-1.61-.48z"
-        fill="url(#prefix__j_dark)"
+        fill={`url(#prefix__j_${svgId})`}
       />
     </svg>
   );